requirements.txt
alembic~=1.13.1
aiohttp~=3.9.2
asyncpg~=0.29.0
asgi-lifespan~=2.1.0
black~=24.4.2
coverage~=7.5.3
dynaconf~=3.2.5
flake8~=7.0.0
fastapi~=0.111.0
httpx~=0.27.0
isort~=5.13.2
loguru~=0.7.2
mypy~=1.10.0
pydantic~=2.7.3
pyhumps~=3.8.0
pytz~=2024.1
pytest~=8.2.2
pytest-asyncio~=0.23.7
python-keycloak~=4.1.0
python-jose~=3.3.0
pytest-mock~=3.14.0
python-dotenv~=1.0.1
requests~=2.32.3
starlette~=0.37.2
SQLAlchemy~=2.0.30
tronpy~=0.5.0
uvicorn~=0.30.1
